remove label from textfield
